ValueController
+ FINALLY! After hours of frustration, Values now get created as Value object (Relationships should work, at least I hope so!). Functionality now is guaranteed although there still is much to do
+ Value-Migrations, Links and Creation
+ Generation (without validation, not necesarry at this point bcs everything IS already validated before) as static method in Value.php

+++ database/migrations/2019_08_12_132416_create_values_table.php
@@ -16,11 +16,11 @@ class CreateValuesTable extends Migration
         Schema::create('values', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->decimal('value');
-            $table->integer('discipline_id');
-            $table->integer('student_id');
+            $table->unsignedBigInteger('discipline_id');
+            $table->unsignedBigInteger('student_id');
             $table->string('class');
             $table->string('year');
-            $table->integer('teacher_id');
+            $table->unsignedBigInteger('teacher_id');
             $table->dateTime('datetime');
             $table->timestamps();
         });
